+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+## [0.147.6] - 2026-07-23
+
+### Changes
+- fix(shifts): show full shift notes on browse cards ([#1147](https://github.com/everybody-eats-nz/volunteer-portal/pull/1147)) by @malinmalliyawadu
+
+
+## [0.147.5] - 2026-07-23
+
+### Changes
+- fix: load newly created locations immediately on create-shift page ([#1143](https://github.com/everybody-eats-nz/volunteer-portal/pull/1143)) by @malinmalliyawadu
+
+
+## [0.147.4] - 2026-07-23
+
+### Changes
+- fix(admin): include selected locations in users page location filter ([#1145](https://github.com/everybody-eats-nz/volunteer-portal/pull/1145)) by @malinmalliyawadu
+
+
 ## [0.147.3] - 2026-07-23
 
 ### Changes